change to take care of warning msg
git-svn-id: svn://irodssvn.ucsd.edu/trunk@117 ea137bd6-aa86-47b6-b90d-b101d9f19a90

+++ iRODS/server/api/src/rsRuleExecSubmit.c
@@ -123,7 +123,7 @@ getReiFilePath (char *reiFilePath, char *userName)
 
     snprintf (reiFilePath, MAX_NAME_LEN,
      "%-s/%-s/%-s.%-s.%-d", getConfigDir(), PACKED_REI_DIR,
-       REI_FILE_NAME, myUserName, random());
+       REI_FILE_NAME, myUserName, (uint) random());
 
     return (0);
 }
